# openapi.yaml
openapi: 3.1.0
info:
  title: Pipeline Labs SDK API
  version: 1.0.0
  description: |
    Curated SDK surface for Stainless. This spec is intentionally limited to the
    most relevant programmatic routes for dataset ingestion, AI processing, job
    tracking, downloads, chat, auth verification, billing, and health.
servers:
  - url: https://pipeline-labs.vercel.app
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/upload:
    post:
      operationId: uploadDataset
      tags: [Upload]
      summary: Upload or finalize a dataset
      description: |
        Single upload endpoint supporting:
        - simple multipart upload
        - chunked upload init
        - chunk upload
        - chunked upload finalize
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                sessionId:
                  type: string
                chunkIndex:
                  type: integer
                totalChunks:
                  type: integer
                chunk:
                  type: string
                  format: binary
                filename:
                  type: string
                fileType:
                  type: string
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ChunkInitRequest'
                - $ref: '#/components/schemas/ChunkFinalizeRequest'
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/upload/chunk:
    post:
      operationId: uploadDatasetChunk
      tags: [Upload]
      summary: Compatibility alias for chunked uploads
      description: Thin compatibility route that forwards to `/api/upload`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                chunkIndex:
                  type: integer
                totalChunks:
                  type: integer
                chunk:
                  type: string
                  format: binary
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ChunkInitRequest'
                - $ref: '#/components/schemas/ChunkFinalizeRequest'
      responses:
        '200':
          description: Upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    ChunkInitRequest:
      type: object
      required: [action, filename, fileSize, fileType, totalChunks]
      properties:
        action: { type: string, enum: [init] }
        filename: { type: string }
        fileSize: { type: integer }
        fileType: { type: string, enum: [csv, xlsx, xls, json] }
        totalChunks: { type: integer }
    ChunkFinalizeRequest:
      type: object
      required: [action, sessionId]
      properties:
        action: { type: string, enum: [finalize] }
        sessionId: { type: string }
    UploadResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        path: { type: string }
        publicUrl: { type: string, format: uri }
        name: { type: string }
        size: { type: integer }
        type: { type: string }
        rowCount: { type: integer }
        columnCount: { type: integer }
        createdAt: { type: string, format: date-time }
        sessionId: { type: string }
        chunkSize: { type: integer }
        totalChunks: { type: integer }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }